pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-skill-server"
version = "0.1.0"
description = "Develop AI skills locally in your editor, deploy to production when ready"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Janine Cheng" },
]
keywords = ["mcp", "skills", "llm", "agents", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gcs = [
    "google-cloud-storage>=2.0.0",
]

[project.scripts]
mcp-skill-server = "mcp_skill_server.cli:main"
mcp-skill-init = "mcp_skill_server.cli:main_init"
mcp-skill-validate = "mcp_skill_server.cli:main_validate"

[project.urls]
Homepage = "https://github.com/jcc-ne/mcp-skill-server"
Repository = "https://github.com/jcc-ne/mcp-skill-server"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_skill_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "--cov=mcp_skill_server --cov-report=term-missing"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
]
